MOMO-28 feat: apply redux state (17)
RoutineAddListComponent 로 분리 후 적용
