+++ src/screens/tutorials/MainTutorialScreen.js
@@ -6,7 +6,6 @@ import { InitTutorialScreen } from './InitTutorialScreen';
 import { TimePickerScreen } from './TimePickerScreen';
 import { ButtonBottom } from '../../components/Buttons/ButtonBottom';
 import { Step } from '../../utils/tutorials/Step';
-import { RoutineAddList } from '../../components/RoutineAddList/RoutineAddListModal';
 
 import { useDispatch, useSelector } from 'react-redux';
 import { 
@@ -71,7 +70,7 @@ export const MainTutorialScreen = () => {
                 <View style={{flex: 1}}>
                     {
                         !isStepScreen ? <InitTutorialScreen/> : 
-                        <Animated.View style={{opacity: opacityAnimation}}>
+                        <Animated.View style={{opacity: opacityAnimation, flex:1}}>
                             <TimePickerScreen/>
                         </Animated.View>
                     }
